+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 * added PharExtractor
 * added BinaryProvider
 * added NotifuDriver (can be used on Windows Seven)
+* added ToasterDriver (can be used on Windows Eight and higher)